pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uaef"
version = "0.1.0"
description = "Universal Autonomous Enterprise Fabric - Coordinate autonomous agents with trust verification and automated settlement"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "UAEF Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.12",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",

    # API
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "httpx>=0.25",

    # Agent SDK
    "anthropic>=0.39",

    # Security
    "cryptography>=41.0",
    "pyjwt>=2.8",
    "passlib>=1.7",

    # Utilities
    "structlog>=23.2",
    "python-json-logger>=2.0",
    "tenacity>=8.2",
    "python-dotenv>=1.0",

    # AWS (serverless)
    "boto3>=1.34",
    "aws-lambda-powertools>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "aiosqlite>=0.19",
    "mypy>=1.7",
    "ruff>=0.1",
    "black>=23.11",
    "pre-commit>=3.6",
    "faker>=21.0",
]
azure = [
    "azure-functions>=1.17",
    "azure-identity>=1.15",
    "azure-servicebus>=7.11",
]

[project.urls]
Homepage = "https://github.com/uaef/uaef"
Documentation = "https://github.com/uaef/uaef/docs"

[tool.hatch.build.targets.wheel]
packages = ["src/uaef"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=uaef --cov-report=term-missing"
